+++ package_monitor/management/commands/packagemonitorcli.py
@@ -1,4 +1,4 @@
-"""CLI commands for Package Monitor."""
+"""CLI tool for Package Monitor."""
 
 import datetime as dt
 import json
@@ -21,6 +21,10 @@ except ImportError:
     yaml = None
 
 COMMAND = "command"
+
+FORMAT_YAML = "yaml"
+FORMAT_JSON = "json"
+FORMAT_DEFAULT = FORMAT_YAML
 
 EXCLUDED_METADATA_FIELDS = {"description", "description_content_type", "license"}
 """Metadata fields excluded by default in the dump command."""
@@ -51,8 +55,8 @@ class Command(BaseCommand):
         dump.add_argument(
             "-f",
             "--format",
-            default="yaml",
-            choices=["json", "yaml"],
+            default=FORMAT_DEFAULT,
+            choices=[FORMAT_JSON, FORMAT_YAML],
             help="Data format. (Default: yaml)",
         )
         dump.add_argument(
@@ -74,12 +78,18 @@ class Command(BaseCommand):
             action="store_true",
             help="Resolve paths for all files",
         )
-        subparsers.add_parser(
+        install = subparsers.add_parser(
             SubCommand.INSTALL.value,
             help=(
                 "Print parameters for installing outdated packages. "
                 "Usage: pip install $(python manage.py packagemonitorcli install)"
             ),
+        )
+        install.add_argument(
+            "-r",
+            "--refresh",
+            action="store_true",
+            help="When set will first refresh the list of distributions pages",
         )
         subparsers.add_parser(
             SubCommand.OUTDATED.value,
@@ -95,36 +105,28 @@ class Command(BaseCommand):
         )
 
     def handle(self, *args, **options):
+        functions = {
+            SubCommand.DUMP: self.dump,
+            SubCommand.REFRESH: self.refresh,
+            SubCommand.INSTALL: self.install,
+            SubCommand.OUTDATED: self.outdated,
+            SubCommand.VERSION: self.version,
+        }
         command = options[COMMAND]
-        if command == SubCommand.DUMP:
-            self.dump(
-                format=options["format"],
-                all_meta_fields=options["all_meta_fields"],
-                all_import_paths=options["all_import_paths"],
-                resolve_files=options["resolve_files"],
-            )
-        elif command == SubCommand.REFRESH:
-            self.refresh()
-        elif command == SubCommand.INSTALL:
-            self.install()
-        elif command == SubCommand.OUTDATED:
-            self.outdated()
-        elif command == SubCommand.VERSION:
-            self.version()
-        else:
-            raise NotImplementedError(command)
+        try:
+            f = functions[command]
+        except KeyError:
+            raise NotImplementedError(command) from None
+        f(**options)
 
-    def dump(
-        self,
-        format: str,
-        all_meta_fields: bool,
-        resolve_files: bool,
-        all_import_paths: bool,
-    ):
-        if all_import_paths:
+    def dump(self, **options: dict):
+        if options.get("all_import_paths"):
             import_paths = sys.path
         else:
             import_paths = distribution_packages.relevant_import_paths()
+
+        all_meta_fields = bool(options.get("all_meta_fields"))
+        resolve_files = bool(options.get("resolve_files"))
         duplicate_count = 0
         packages = {}
         for d in importlib_metadata.distributions(path=import_paths):
@@ -179,16 +181,16 @@ class Command(BaseCommand):
             },
             "packages": packages,
         }
-
-        if format == "json":
+        output_format = options.get("format", FORMAT_DEFAULT)
+        if output_format == FORMAT_JSON:
             o = json.dumps(data, sort_keys=True, indent=4)
-        elif format == "yaml":
+        elif output_format == FORMAT_YAML:
             o = yaml.dump(data)
         else:
-            raise NotImplementedError(format)
+            raise NotImplementedError(output_format)
         self.stdout.write(o)
 
-    def refresh(self):
+    def refresh(self, **_):
         package_count = Distribution.objects.count()
         self.stdout.write(
             f"Started refreshing data for currently {package_count} distribution packages...."
@@ -200,7 +202,7 @@ class Command(BaseCommand):
         )
         self.stdout.write(self.style.SUCCESS("DONE"))
 
-    def outdated(self):
+    def outdated(self, **_):
         obj = Distribution.objects.first()
         updated_at = obj.updated_at if obj else None
         outdated_count = (
@@ -212,7 +214,7 @@ class Command(BaseCommand):
         last_checked = humanize.naturaltime(updated_at) if updated_at else "?"
         self.stdout.write(
             f"{outdated_count} / {package_count} distribution packages are outdated. "
-            f"Last checked {last_checked}"
+            f"({last_checked})"
         )
         if outdated_count == 0:
             return
@@ -229,7 +231,10 @@ class Command(BaseCommand):
                 f"{d.name} {d.latest_version} [upgradeable from: {d.latest_version}]"
             )
 
-    def install(self):
+    def install(self, **options):
+        refresh = options.get("refresh")
+        if refresh:
+            Distribution.objects.update_all()
         outdated_count = (
             Distribution.objects.filter_visible()
             .exclude(latest_version="")
@@ -247,5 +252,5 @@ class Command(BaseCommand):
         )
         self.stdout.write(command)
 
-    def version(self):
+    def version(self, **_):
         self.stdout.write(__version__)